backend/requirements.txt
affine==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
APScheduler==3.11.1
asttokens==3.0.1
attrs==25.4.0
bidict==0.23.1
cattrs==25.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dnspython==2.8.0
docutils==0.22.3
email-validator==2.3.0
executing==2.2.1
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
flatbuffers==25.9.23
fonttools==4.60.1
freezegun==1.5.5
frozenlist==1.8.0
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
greenlet==3.2.4
h11==0.16.0
holidays==0.85
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jh2==5.0.10
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mercantile==1.2.1
multidict==6.7.0
narwhals==2.12.0
nest-asyncio==1.6.0
nicegui==3.3.1
niquests==3.15.2
numpy==2.3.5
openmeteo_requests==1.7.4
openmeteo_sdk==1.23.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyodbc==5.3.0
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.15.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qh3==1.5.6
requests==2.32.5
requests-cache==1.2.1
retry-requests==2.0.0
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentry-sdk==2.46.0
shapely==2.1.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
url-normalize==2.2.1
urllib3==2.5.0
urllib3-future==2.14.907
uvicorn==0.38.0
uvloop==0.22.1
wassima==2.0.2
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wsproto==1.3.2
xgboost==3.1.2
xyzservices==2025.11.0
yarl==1.22.0
